refactor: 폴더 구조, 절대경로 수정, appshell 삭제
